Implement DateFilter
In order to make query on date/datetime type, I implement the AbstractDateFilter.
For example if you have a date in the ISO 8601 date format :
->add('date', 'sonata_search_elastica_datetime', null, 'datetime', array(
            'format' => 'c'
        ))
The format must be a string formatted according to the php format date.

You can also define a date range with a sonata_search_elastica_date_range